+++ NEWS.md
@@ -2,3 +2,5 @@
 
 * Set up basic package components.
 * Created `lr_check()` function.
+* Added check and coverage GitHub actions.
+* Added simple tests for `lr_check()`.